feat(collab): add lobby component
